requirements.txt
asgiref==3.4.1
cryptography==3.4.8
dj-database-url==0.5.0
Django==3.1.13
django-cors-headers==3.8.0
django-heroku==0.3.1
django-rest-framework==0.1.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.3.1
psycopg2==2.9.1
pycparser==2.20
python-dotenv==0.19.2
whitenoise==5.3.0
